fixed naming issues
deleted unused routes and classes
