--- lib/pages/info.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Info extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Info'),
-        backgroundColor: Theme.of(context).primaryColor,
-      ),
-    );
-  }
-}
