+++ lib/pages/route/route.dart
@@ -1,15 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:hiking4nerds/components/navbar.dart';
 
-class Plan extends StatelessWidget {
+class Route extends StatefulWidget {
+  @override
+  _RouteState createState() => _RouteState();
+}
+
+class _RouteState extends State<Route> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Plan'),
+        title: Text('Route'),
         backgroundColor: Theme.of(context).primaryColor,
       ),
-      bottomNavigationBar: Navbar(),
+      bottomNavigationBar: NavBar(),
     );
   }
 }
