--- lib/pages/settings.dart
@@ -1,20 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:hiking4nerds/components/navbar.dart';
-
-class Settings extends StatefulWidget {
-  @override
-  _SettingsState createState() => _SettingsState();
-}
-
-class _SettingsState extends State<Settings> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Settings'),
-        backgroundColor: Theme.of(context).primaryColor,
-      ),
-      bottomNavigationBar: Navbar(),
-    );
-  }
-}
